+++ Assets/Scripts/Stored/Antiquity.cs
@@ -8,9 +8,9 @@ namespace Stored
         [SerializeField] private string displayName;
         [SerializeField] private Sprite sprite;
         [Tooltip("The health at which this will start taking damage")]
-        [SerializeField] private int breakingHealth;
+        [SerializeField] private float breakingHealth;
         [Tooltip("The maximum damage this can take, as such the total health")]
-        [SerializeField] private int maxHealth;
+        [SerializeField] private float maxHealth;
         [Tooltip("THe final score you get after cleaning the fossil")]
         [SerializeField] private float score;
         [Tooltip("Set to true to manually choose its set. Otherwise it'll automatically be assigned a set if a set" +
@@ -19,8 +19,8 @@ namespace Stored
 
         public string DisplayName => displayName;
         public Sprite Sprite => sprite;
-        public int BreakingHealth => breakingHealth;
-        public int MaxHealth => maxHealth;
+        public float BreakingHealth => breakingHealth;
+        public float MaxHealth => maxHealth;
         public bool OverrideSet => overrideSet;
         public float Score => score;
         public AntiquitySet AntiquitySet;
